# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "단체급식 프로모션 신청 서비스 API",
    "version": "1.0.0",
    "description": "영양사(MANAGER)가 프로모션을 등록하고 취식자(PARTICIPANT)가 조회/신청하는 단체급식 프로모션 신청 서비스 API. 인증은 JWT access token(짧은 만료) + refresh token(긴 만료, 무상태)을 사용한다."
  },
  "servers": [
    {
      "url": "http://localhost:{port}/api",
      "description": "로컬 개발 서버 (프론트엔드와 별도 오리진, CORS로 접근. 포트는 .env의 PORT 값)",
      "variables": {
        "port": { "default": "3000" }
      }
    }
  ],
  "tags": [
    { "name": "Auth", "description": "회원가입/로그인/토큰 재발급" },
    { "name": "Promotions", "description": "프로모션 등록/조회/수정/삭제" },
    { "name": "Applications", "description": "참가 신청/취소/신청 현황" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "로그인 응답으로 받은 accessToken을 사용한다."
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": ["MANAGER", "PARTICIPANT"],
        "description": "MANAGER: 영양사, PARTICIPANT: 취식자"
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "홍길동" },
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "role": { "$ref": "#/components/schemas/Role" }
        },
        "required": ["id", "name", "email", "role"]
      },
      "Promotion": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "managerId": { "type": "integer", "example": 1 },
          "title": { "type": "string", "example": "여름맞이 특식 이벤트" },
          "description": { "type": "string", "example": "냉면 시식 프로모션입니다." },
          "applyStartAt": { "type": "string", "format": "date", "example": "2026-08-14" },
          "applyEndAt": { "type": "string", "format": "date", "example": "2026-08-20" },
          "eventDate": { "type": "string", "format": "date", "example": "2026-08-21" },
          "capacity": { "type": "integer", "example": 50 },
          "appliedCount": { "type": "integer", "example": 10 }
        },
        "required": ["id", "managerId", "title", "description", "applyStartAt", "applyEndAt", "eventDate", "capacity", "appliedCount"]
      },
      "Application": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "promotionId": { "type": "integer", "example": 1 },
          "userId": { "type": "integer", "example": 5 },
          "appliedAt": { "type": "string", "format": "date-time", "example": "2026-08-14T09:00:00.000Z" }
        },
        "required": ["id", "promotionId", "userId", "appliedAt"]
      },
      "Applicant": {
        "type": "object",
        "properties": {
          "userId": { "type": "integer", "example": 5 },
          "name": { "type": "string", "example": "김취식" },
          "appliedAt": { "type": "string", "format": "date-time", "example": "2026-08-14T09:00:00.000Z" }
        },
        "required": ["userId", "name", "appliedAt"]
      },
      "SignupRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "홍길동" },
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "format": "password", "example": "password123" },
          "role": { "$ref": "#/components/schemas/Role" }
        },
        "required": ["name", "email", "password", "role"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "format": "password", "example": "password123" }
        },
        "required": ["email", "password"]
      },
      "LoginResponse": {
        "type": "object",
        "description": "user 정보는 내려주지 않는다. 클라이언트는 accessToken(JWT)의 payload({id, role})를 디코드해 역할을 판단한다.",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" }
        },
        "required": ["accessToken", "refreshToken"]
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": { "type": "string" }
        },
        "required": ["refreshToken"]
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" }
        },
        "required": ["accessToken"]
      },
      "PromotionCreateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "여름맞이 특식 이벤트" },
          "description": { "type": "string", "example": "냉면 시식 프로모션입니다." },
          "applyStartAt": { "type": "string", "format": "date", "example": "2026-08-14" },
          "applyEndAt": { "type": "string", "format": "date", "example": "2026-08-20" },
          "eventDate": { "type": "string", "format": "date", "example": "2026-08-21" },
          "capacity": { "type": "integer", "minimum": 1, "example": 50 }
        },
        "required": ["title", "description", "applyStartAt", "applyEndAt", "eventDate", "capacity"]
      },
      "PromotionUpdateRequest": {
        "type": "object",
        "description": "등록자 본인만 수정 가능. 전달된 필드만 갱신한다.",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "applyStartAt": { "type": "string", "format": "date" },
          "applyEndAt": { "type": "string", "format": "date" },
          "eventDate": { "type": "string", "format": "date" },
          "capacity": { "type": "integer", "minimum": 1 }
        }
      },
      "MyApplicationStatus": {
        "type": "object",
        "description": "로그인한 취식자 본인이 해당 프로모션에 신청했는지 여부.",
        "properties": {
          "applied": { "type": "boolean", "example": true },
          "appliedAt": { "type": "string", "format": "date-time", "nullable": true, "example": "2026-08-14T09:00:00.000Z" }
        },
        "required": ["applied", "appliedAt"]
      },
      "ApplicationStatusResponse": {
        "type": "object",
        "properties": {
          "promotionId": { "type": "integer", "example": 1 },
          "title": { "type": "string", "example": "여름맞이 특식 이벤트" },
          "capacity": { "type": "integer", "example": 50 },
          "appliedCount": { "type": "integer", "example": 10 },
          "applicants": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Applicant" }
          }
        },
        "required": ["promotionId", "title", "capacity", "appliedCount", "applicants"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "요청을 처리할 수 없습니다." }
        },
        "required": ["message"]
      }
    },
    "parameters": {
      "PromotionId": {
        "name": "promotionId",
        "in": "path",
        "required": true,
        "description": "프로모션 ID",
        "schema": { "type": "integer" }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "요청 값이 유효하지 않음",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        }
      },
      "Unauthorized": {
        "description": "인증되지 않음 (토큰 없음/위조/만료)",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        }
      },
      "Forbidden": {
        "description": "권한 없음",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        }
      },
      "NotFound": {
        "description": "리소스를 찾을 수 없음",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        }
      },
      "Conflict": {
        "description": "리소스 상태 충돌 (중복 등)",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "회원가입",
        "description": "이름, 이메일, 비밀번호, 역할(MANAGER/PARTICIPANT)로 회원가입한다. 이메일은 중복 가입할 수 없다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SignupRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/User" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": {
            "description": "이미 사용 중인 이메일",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "로그인",
        "description": "이메일/비밀번호로 로그인하여 access token과 refresh token을 발급받는다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } }
            }
          },
          "401": {
            "description": "이메일 또는 비밀번호 불일치",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "액세스 토큰 재발급",
        "description": "refresh token(무상태, 서명 검증만 수행)으로 새 access token을 발급받는다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "재발급 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RefreshResponse" } }
            }
          },
          "401": {
            "description": "refresh token이 없거나 위조/만료됨",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/promotions": {
      "post": {
        "tags": ["Promotions"],
        "summary": "프로모션 등록 (영양사 전용)",
        "description": "MANAGER만 등록 가능. 진행일(eventDate)은 신청 종료일(applyEndAt)과 같거나 이후여야 한다. appliedCount는 0으로 시작한다.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PromotionCreateRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "등록 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Promotion" } }
            }
          },
          "400": {
            "description": "필수 항목 누락 또는 진행일이 신청 종료일보다 이전",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "PARTICIPANT는 등록 불가",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      },
      "get": {
        "tags": ["Promotions"],
        "summary": "프로모션 목록 조회",
        "description": "로그인한 사용자(영양사·취식자 공통)가 등록된 프로모션 목록을 조회한다. 모집 인원과 현재 신청 인원(appliedCount)을 포함한다.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "조회 성공 (목록이 없으면 빈 배열)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Promotion" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/promotions/{promotionId}": {
      "get": {
        "tags": ["Promotions"],
        "summary": "프로모션 상세 조회",
        "description": "로그인한 사용자(영양사·취식자 공통)가 프로모션 상세 정보를 조회한다.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PromotionId" }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Promotion" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Promotions"],
        "summary": "프로모션 수정 (등록자 본인만)",
        "description": "등록자 본인만 수정 가능. 신청자가 있으면 모집 인원(capacity)을 현재 신청 인원(appliedCount) 미만으로 축소할 수 없다. 진행일 ≥ 신청 종료일 검증이 수정 시에도 적용된다.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PromotionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PromotionUpdateRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Promotion" } }
            }
          },
          "400": {
            "description": "진행일 제약 위반 또는 신청 인원 미만으로 정원 축소 시도",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "등록자 본인이 아님",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Promotions"],
        "summary": "프로모션 삭제 (등록자 본인만)",
        "description": "등록자 본인만 삭제 가능. 삭제 시 해당 프로모션의 신청 기록도 함께 삭제된다(cascade).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PromotionId" }
        ],
        "responses": {
          "204": { "description": "삭제 성공" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "등록자 본인이 아님",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/promotions/{promotionId}/applications": {
      "post": {
        "tags": ["Applications"],
        "summary": "프로모션 참가 신청 (취식자 전용)",
        "description": "PARTICIPANT만 신청 가능. 신청자는 로그인한 사용자(req.user) 기준으로 기록된다. 신청 기간(applyStartAt~applyEndAt) 내에만 신청 가능하며, 동일 사용자·동일 프로모션 중복 신청과 정원 초과는 서버에서 원자적으로 거부한다.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PromotionId" }
        ],
        "responses": {
          "201": {
            "description": "신청 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Application" } }
            }
          },
          "400": {
            "description": "신청 기간이 아니거나 정원이 마감됨",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "MANAGER는 신청 불가",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "이미 신청한 프로모션에 재신청",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      },
      "get": {
        "tags": ["Applications"],
        "summary": "신청 현황 및 신청자 목록 조회 (영양사 전용, 본인 등록 프로모션만)",
        "description": "MANAGER가 본인이 등록한 프로모션의 현재 신청 인원과 신청자 목록(이름, 신청 일시)을 조회한다. 신청자가 없으면 빈 배열을 반환한다.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PromotionId" }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApplicationStatusResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "본인이 등록한 프로모션이 아님",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/promotions/{promotionId}/applications/me": {
      "get": {
        "tags": ["Applications"],
        "summary": "내 신청 여부 조회 (취식자 전용)",
        "description": "로그인한 취식자 본인이 해당 프로모션에 신청했는지, 신청했다면 신청 일시를 조회한다.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PromotionId" }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MyApplicationStatus" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "MANAGER는 조회 불가",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Applications"],
        "summary": "참가 신청 취소 (본인 신청만)",
        "description": "로그인한 취식자 본인의 신청을 취소한다. 본인이 신청하지 않은 프로모션에 대한 취소 요청은 거부된다.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PromotionId" }
        ],
        "responses": {
          "204": { "description": "취소 성공" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "본인의 신청 내역이 아님(신청 내역 없음)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  }
}